% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_HTML_functions.R
\name{check_cosubject}
\alias{check_cosubject}
\title{Returns all cosubjects listed in this biography}
\usage{
check_cosubject(htmltext)
}
\arguments{
\item{htmltext}{Character vector : only section pertaining to biography}
}
\value{
vector of cosubjects in biography
}
\description{
Returns all cosubjects listed in this biography
}

