% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NER_improve.R
\name{check_entity_in_fullset}
\alias{check_entity_in_fullset}
\title{<<BasicInfo>>}
\usage{
check_entity_in_fullset(entity)
}
\arguments{
\item{entity}{temp}
}
\value{
temp
}
\description{
<<BasicInfo>>
}

