% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NER_processing.R
\name{combine_two}
\alias{combine_two}
\title{Takes two predictor vectors and outputs a combined predictor vector, giving priority to vector a.}
\usage{
combine_two(a, b)
}
\arguments{
\item{a}{tagged output}

\item{b}{tagged output}
}
\value{
combined tagged output
}
\description{
Takes two predictor vectors and outputs a combined predictor vector, giving priority to vector a.
}

