% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NER_processing.R
\name{compute_parenthesis}
\alias{compute_parenthesis}
\title{This function computes the level of parenthesis embedding}
\usage{
compute_parenthesis(vec)
}
\arguments{
\item{vec}{text}
}
\value{
vector of parenthesis embedding
}
\description{
This function computes the level of parenthesis embedding
}

