% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_HTML_functions.R
\name{dnb_grab_main}
\alias{dnb_grab_main}
\title{This extracts the pertinent part of the html code and outputs that.}
\usage{
dnb_grab_main(raw_htmltext)
}
\arguments{
\item{raw_htmltext}{Character vector corresponding to lines in a document (HTML code)}
}
\value{
Character vector contains biography section of code
}
\description{
This extracts the pertinent part of the html code and outputs that.
}

