% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_HTML_functions.R
\name{exists_cosubject}
\alias{exists_cosubject}
\title{Detects if this there are cosubject mentions in this biography}
\usage{
exists_cosubject(htmltext)
}
\arguments{
\item{htmltext}{Character vector : only section pertaining to biography}
}
\value{
T/F: True if 'cosubject' appears in document'
}
\description{
Detects if this there are cosubject mentions in this biography
}

