% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NER_improve.R
\name{find_closest_date}
\alias{find_closest_date}
\title{<<BasicInfo>>}
\usage{
find_closest_date(matches, tag_df, doc_split = 500)
}
\arguments{
\item{matches}{temp}

\item{tag_df}{temp}

\item{doc_split}{temp}
}
\value{
temp
}
\description{
<<BasicInfo>>
}

