% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NER_processing.R
\name{format_text_and_split}
\alias{format_text_and_split}
\title{This function does the following:
1. add spaces before punctuation
2. add spaces before/after (), [], {}, as to separate them off => easier to count
3. remove more than one space in a row.
4. Tokenize (split into a word-by-word representation).
4a. Fix problems with ;.}
\usage{
format_text_and_split(text)
}
\arguments{
\item{text}{character vector}
}
\value{
formatted text
}
\description{
This function does the following:
1. add spaces before punctuation
2. add spaces before/after (), [], {}, as to separate them off => easier to count
3. remove more than one space in a row.
4. Tokenize (split into a word-by-word representation).
4a. Fix problems with ;.
}

