% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NER_improve.R
\name{improve_pred}
\alias{improve_pred}
\title{<<BasicInfo>>}
\usage{
improve_pred(comb.tag, main.person = NULL, exact.words = NULL,
  regex.words = NULL)
}
\arguments{
\item{comb.tag}{temp}

\item{main.person}{temp}

\item{exact.words}{temp}

\item{regex.words}{temp}
}
\value{
temp
}
\description{
<<BasicInfo>>
}

