% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NER_improve.R
\name{insert_tag}
\alias{insert_tag}
\title{<<BasicInfo>>}
\usage{
insert_tag(start, len, tags)
}
\arguments{
\item{start}{temp}

\item{len}{temp}

\item{tags}{temp}
}
\value{
temp
}
\description{
<<BasicInfo>>
}

