% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_HTML_functions.R
\name{is_nobio}
\alias{is_nobio}
\title{Checks if the HTML file is a biography}
\usage{
is_nobio(raw_htmltext)
}
\arguments{
\item{raw_htmltext}{Character vector corresponding to lines in a document (HTML code)}
}
\value{
TRUE if file is not a biography
}
\description{
Checks if the HTML file is a biography
}

