% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NER_processing.R
\name{normalize_tagtext}
\alias{normalize_tagtext}
\title{This vectorizes and normalizes the tagged text by standardizing the tags to have the same format between taggers.}
\usage{
normalize_tagtext(text, true.text, type)
}
\arguments{
\item{text}{tagged text}

\item{true.text}{raw text}

\item{type}{"ST" or "LP"}
}
\value{
normalized form of tagged text
}
\description{
This vectorizes and normalizes the tagged text by standardizing the tags to have the same format between taggers.
}

