% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_HTML_functions.R
\name{ODNB_extract_name_segment}
\alias{ODNB_extract_name_segment}
\title{This function returns a list of two elements: the substring in the text that matches the name segment
(including html), and the character number that signifies the end of the string.}
\usage{
ODNB_extract_name_segment(text)
}
\arguments{
\item{text}{odnb text}
}
\value{
FIX THIS
}
\description{
This function returns a list of two elements: the substring in the text that matches the name segment
(including html), and the character number that signifies the end of the string.
}

