% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_HTML_functions.R
\name{ODNB_fix_accent_html}
\alias{ODNB_fix_accent_html}
\title{This removes/fixes all sorts of HTML things.}
\usage{
ODNB_fix_accent_html(text)
}
\arguments{
\item{text}{char vec (ODNB)}
}
\value{
char vec
}
\description{
This removes/fixes all sorts of HTML things.
}

