% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NER_processing.R
\name{paste_two}
\alias{paste_two}
\title{this pastes together two consecutive entries in a vector
starting with ind.start and then ind.start+1
and edits the vector to remove the extra cell}
\usage{
paste_two(vec, ind.start)
}
\arguments{
\item{vec}{vector}

\item{ind.start}{where to paste}
}
\value{
updated vector
}
\description{
this pastes together two consecutive entries in a vector
starting with ind.start and then ind.start+1
and edits the vector to remove the extra cell
}

