% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NER_processing.R
\name{proc_tagtext}
\alias{proc_tagtext}
\title{returns vectors of tags matching tag. PERSON, ORGANIZATION, LOCATION}
\usage{
proc_tagtext(tagged.text, type, tag.types = c("PERSON", "ORGANIZATION",
  "LOCATION"))
}
\arguments{
\item{tagged.text}{vectorized tagged text}

\item{type}{ST or LP}

\item{tag.types}{existing tag.types}
}
\value{
fix this
}
\description{
returns vectors of tags matching tag. PERSON, ORGANIZATION, LOCATION
}

