% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_HTML_functions.R
\name{process_cosubject}
\alias{process_cosubject}
\title{Split cosubject biographies into portions}
\usage{
process_cosubject(main_id, all_ids)
}
\arguments{
\item{main_id}{ID number for person the document is mainly about (ignored if group bio)}

\item{all_ids}{Vector of numbers that refer to the same document}
}
\value{
list of cosubject biographies
}
\description{
This function determines the type of cosubject biography,
and consequently splits it (and returns a list of texts and id numbers for the texts)
}
\details{
REQUIRES ODNB_rawHTML to exist in global environment.
}

