% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NER_improve.R
\name{remove_badentities}
\alias{remove_badentities}
\title{<<BasicInfo>>}
\usage{
remove_badentities(comb.tag.mat, position.df)
}
\arguments{
\item{comb.tag.mat}{temp}

\item{position.df}{temp}
}
\value{
temp
}
\description{
<<BasicInfo>>
}

