% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NER_improve.R
\name{remove_entries}
\alias{remove_entries}
\title{<<BasicInfo>>}
\usage{
remove_entries(tags, entries)
}
\arguments{
\item{tags}{temp}

\item{entries}{temp}
}
\value{
temp
}
\description{
<<BasicInfo>>
}

