% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_HTML_functions.R
\name{rm_emptylines}
\alias{rm_emptylines}
\title{Removes all elements which only have spaces}
\usage{
rm_emptylines(raw_htmltext)
}
\arguments{
\item{raw_htmltext}{Character vector corresponding to lines in a document (HTML code)}
}
\value{
Character vector without empty lines
}
\description{
Removes all elements which only have spaces
}

