% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NER_processing.R
\name{tag_dates}
\alias{tag_dates}
\title{Finds dates between 1200-1900}
\usage{
tag_dates(vec, mindate = 1100, maxdate = 1900)
}
\arguments{
\item{vec}{text}

\item{mindate}{temp}

\item{maxdate}{temp}
}
\value{
vector with dates extracted...
}
\description{
Finds dates between 1200-1900
}

